deployment/flask_ai/requirements-k8s.txt
flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
psutil>=5.9.0
numpy>=1.24.0,<2.0.0
transformers>=4.35.0,<5.0.0
huggingface-hub>=0.20.0,<1.0.0

# Observability
prometheus-flask-exporter>=0.23.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-instrumentation-flask>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
